Return information on a link after submission.
Previously, when submitting a link to reddit, you would only get back an error if something went wrong. Now, you can get an NSDictionary containing various pieces of information about the submission.

Instead of just returning a full name or identifier, I chose to return the dictionary provided by reddit, as it also includes the URL of the new post.